4th commit

+++ controllers/cards.js
@@ -2,22 +2,26 @@ const Card = require('../models/card.js');
 
 module.exports.getCards = (req, res) => {
   Card.find({})
-    .then((cards) => res.send({ data: cards }))
-    .catch((err) => res.status(404).send({ message: err.message }));
+    .then((cards) => {
+      if (cards.length !== 0) {
+        res.send({ data: cards });
+      } else res.status(400).send({ message: 'Ресурсы не созданы на сервере' });
+    })
+    .catch((err) => res.status(500).send({ message: err.message }));
 };
 
 module.exports.getCardById = (req, res) => {
   Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        res.status(404).send(`Карточки с id : ${req.params.cardId} не существует!`);
+        res.status(400).send(`Карточки с id : ${req.params.cardId} не существует!`);
       } else {
         res.send({ data: card });
       }
     })
     .catch((err) => {
       if (err.message.indexOf('Cast to ObjectId failed') === 0) {
-        res.status(404).send(`Карточка с id : ${req.params.cardId} не найдена!`);
+        res.status(400).send(`Карточка с id : ${req.params.cardId} не найдена!`);
         return;
       }
       res.status(500).send({ message: err });
@@ -33,17 +37,19 @@ module.exports.createCard = (req, res) => {
 };
 
 module.exports.deleteCardById = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        res.status(404).send(`Карточки с id : ${req.params.cardId} не существует!`);
-      } else {
-        res.send({ data: card });
-      }
+        res.status(400).send(`Карточки с id : ${req.params.cardId} не существует!`);
+      } else if (req.user._id === card.owner.toString()) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then((cardRemove) => res.send({ remove: cardRemove }))
+          .catch((err) => res.status(500).send({ message: err }));
+      } else res.status(403).send({ message: 'Недостаточно прав доступа к ресурсу' });
     })
     .catch((err) => {
       if (err.message.indexOf('Cast to ObjectId failed') === 0) {
-        res.status(404).send(`Карточка с id : ${req.params.cardId} не найдена!`);
+        res.status(400).send(`Карточка с id : ${req.params.cardId} не может быть в природе!`);
         return;
       }
       res.status(500).send({ message: err });
@@ -58,14 +64,14 @@ module.exports.likeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        res.status(404).send(`Карточки с id : ${req.params.cardId} не существует!`);
+        res.status(400).send(`Карточки с id : ${req.params.cardId} не существует!`);
       } else {
         res.send({ data: card });
       }
     })
     .catch((err) => {
       if (err.message.indexOf('Cast to ObjectId failed') === 0) {
-        res.status(404).send(`Карточка с id : ${req.params.cardId} не найдена!`);
+        res.status(400).send(`Карточка с id : ${req.params.cardId} не найдена!`);
         return;
       }
       res.status(500).send({ message: err });
@@ -80,14 +86,14 @@ module.exports.dislikeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        res.status(404).send(`Карточки с id : ${req.params.cardId} не существует!`);
+        res.status(400).send(`Карточки с id : ${req.params.cardId} не существует!`);
       } else {
         res.send({ data: card });
       }
     })
     .catch((err) => {
       if (err.message.indexOf('Cast to ObjectId failed') === 0) {
-        res.status(404).send(`Карточка с id : ${req.params.cardId} не найдена!`);
+        res.status(400).send(`Карточка с id : ${req.params.cardId} не найдена!`);
         return;
       }
       res.status(500).send({ message: err });
